Move executor utils to closer package.
Trading one junk drawer for another, to some extent (I hate code called "util"), but it does seem fair to say at least that none of this code will be used outside of executor implementations.
